backend/requirements.txt
fastapi
uvicorn[standard]
ollama
chromadb
sentence-transformers
sqlalchemy
python-dotenv
pydantic
tiktoken
